requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-filter==1.1.0
django-rest-swagger==2.1.2
django-serialize==1.3.0
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
openapi-codec==1.3.2
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2017.3
requests==2.18.4
Serialize==0.1
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
